+++ src/components/closed-captions/closed-captions.tsx
@@ -11,6 +11,8 @@ import {Button} from '../button';
 import {ButtonControl} from '../button-control';
 import {registerToBottomBar} from '../bottom-bar';
 import {redux} from '../../index';
+import {types} from '../../reducers/settings';
+import { isEnter, isSpace } from "../../utils/key-map";
 
 /**
  * mapping state to props
@@ -83,8 +85,21 @@ const ClosedCaptions = connect(mapStateToProps)(
       const onClick = () => {
         const isCCOn = isCaptionsEnabled();
         props.notifyClick(isCCOn);
-        isCCOn ? player.hideTextTrack() : player.showTextTrack();
+        if (isCCOn) {
+          player.hideTextTrack();
+          redux.useStore().dispatch({type: types.UPDATE_IS_CAPTIONS_ENABLED, isCaptionsEnabled: false});
+        } else {
+          player.showTextTrack();
+          redux.useStore().dispatch({type: types.UPDATE_IS_CAPTIONS_ENABLED, isCaptionsEnabled: true});
+        }
       };
+
+        const onKeyDown = (e: KeyboardEvent): void => {
+          if (isEnter(e) || isSpace(e)) {
+            e.preventDefault();
+            onClick();
+          }
+        };
 
         const shouldRender = !!textTracks?.length && props.showCCButton && !props.openMenuFromCCButton;
         props.onToggle(COMPONENT_NAME, shouldRender);
@@ -101,7 +116,9 @@ const ClosedCaptions = connect(mapStateToProps)(
                   onClick={() => {
                     props.notifyClick(true);
                     player.hideTextTrack();
+                    redux.useStore().dispatch({type: types.UPDATE_IS_CAPTIONS_ENABLED, isCaptionsEnabled: false});
                   }}
+                  onKeyDown={onKeyDown}
                 >
                   <Icon type={IconType.ClosedCaptionsOn} />
                 </Button>
@@ -115,7 +132,9 @@ const ClosedCaptions = connect(mapStateToProps)(
                   onClick={() => {
                     player.showTextTrack();
                     props.notifyClick(false);
+                    redux.useStore().dispatch({type: types.UPDATE_IS_CAPTIONS_ENABLED, isCaptionsEnabled: true});
                   }}
+                  onKeyDown={onKeyDown}
                 >
                   <Icon type={IconType.ClosedCaptionsOff} />
                 </Button>